DistanceTo and test refactoring part 3

+++ rgb.go
@@ -23,7 +23,7 @@ func (rgb RGB) RGBA() (r, g, b, a uint32) {
 
 func (rgb RGB) DistanceTo(c color.Color) float64 {
 	other := RGBModel.Convert(c).(RGB)
-	return math.Sqrt(math.Pow(float64(other.R-rgb.R), 2) + math.Pow(float64(other.G-rgb.G), 2) + math.Pow(float64(other.B-rgb.B), 2))
+	return math.Sqrt(math.Pow(float64(other.R)-float64(rgb.R), 2) + math.Pow(float64(other.G)-float64(rgb.G), 2) + math.Pow(float64(other.B)-float64(rgb.B), 2))
 }
 
 var RGBModel color.Model = color.ModelFunc(rgbModel)
